(feat): must use double for long numbers

+++ Core/Rewards/Sums.php
@@ -56,7 +56,7 @@ class Sums
             error_log($e->getMessage());
         }
         
-        return (int) $rows[0]['balance'];
+        return (double) $rows[0]['balance'];
     }
 
 }
